+++ src/main/java/com/parkmate/reservationservice/reservation/event/reservation/ReservationCreateEvent.java
@@ -20,7 +20,7 @@ public class ReservationCreateEvent {
     private String hostUuid;
     private String userUuid;
     private String vehicleNumber;
-    private double amount;
+    private long amount;
     private ReservationStatus status;
     private LocalDateTime entryTime;
     private LocalDateTime exitTime;
@@ -34,7 +34,7 @@ public class ReservationCreateEvent {
                                    String hostUuid,
                                    String userUuid,
                                    String vehicleNumber,
-                                   double amount,
+                                   long amount,
                                    ReservationStatus status,
                                    LocalDateTime entryTime,
                                    LocalDateTime exitTime) {
